requirements.txt
argparse
Pillow
PyPDF2
docx2pdf
comtypes
